backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
black==24.4.2
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crispy-bootstrap4==2024.1
cryptography==43.0.1
distlib==0.3.8
Django==5.1.3
django-allauth==65.0.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.0.0
django-environ==0.12.0
django-filter==24.3
filelock==3.13.1
flatbuffers==24.3.25
humanfriendly==10.0
idna==3.10
imageio==2.36.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
marshmallow==3.22.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.1
pathspec==0.12.1
pdfkit==1.0.0
pillow==11.0.0
pip==25.0
platformdirs==4.1.0
pooch==1.8.2
protobuf==5.29.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
PyMatting==1.1.13
pyreadline3==3.5.4
python-dotenv==1.0.1
pytz==2024.2
qrcode==8.0
razorpay==1.4.2
referencing==0.35.1
rembg==2.0.61
reportlab==4.2.4
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.14.1
sqlparse==0.5.2
sympy==1.13.3
tifffile==2024.12.12
tqdm==4.67.1
tzdata==2024.2
typing_extensions==4.12.2
urllib3==2.3.0
validate-emai
